requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.6.0
pydantic[email]>=2.0
pydantic-settings>=2.0
bcrypt>=4.0.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
apscheduler>=3.10.0
